Fix threads and read arguments from cli for file names
Reading from cli is a requirement stated in the lab sheet.
As for the threading, I wanted to automate creation as much as
possible since having hard coded threads are not a good approach to
begin with. Also, wanted to learn more in one go, so its just for personal
thing :)

Signed-off-by: Taylan Dogan <doganntaylan@gmail.com>