Add singleton as for global map
Well, there should be a global map for keeping words and their counts efficiently. So I chose adding singleton class which distributes the same hash map to each thread. I simply injected it as a dependency, so each thread is working on the same hash map.

As for the Driver class, I want to clean main as much as I can. Also, main is a static function, which makes declaring class variables harder comparing to the normal classes. With separate driver class, I can declare non-static private variables and use them.

Signed-off-by: Taylan Dogan <doganntaylan@gmail.com>